// src/app/services/models/state-wise.model.ts
export interface Statewise {
    active?: string;
    confirmed?: string;
    deaths?: string;
    deltaconfirmed?: string;
    deltadeaths?: string;
    deltarecovered?: string;
    deltaactive?: string;
    lastupdatedtime?: string;
    recovered?: string;
    state?: string;
    statecode: string;
}